Updated `godot-cpp` to 4.6.0

+++ src/audio_stream_opus_chunked.cpp
@@ -327,7 +327,7 @@ PackedVector2Array AudioStreamOpusChunked::pop_front_chunk(int frames) {
 }
 
 
-double AudioStreamPlaybackOpusChunked::_get_stream_sampling_rate() const { 
+float AudioStreamPlaybackOpusChunked::_get_stream_sampling_rate() const { 
     return base->mix_rate; 
 };
